requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2.1
django-ckeditor==6.7.1
django-cuser==2017.3.16
django-environ==0.11.2
django-js-asset==2.2.0
django-modeltranslation==0.19.4
django-phonenumber-field==8.0.0
django-tinymce==4.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.1
executing==2.0.1
frozenlist==1.4.1
idna==3.7
jedi==0.19.1
magic-filter==1.0.12
matplotlib-inline==0.1.7
multidict==6.0.5
parso==0.8.4
phonenumbers==8.13.39
pillow==10.3.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
yarl==1.9.4
psycopg2-binary==2.9.9
sentry-sdk==2.9.0
asgiref==3.8.1
aiogram==3.10.0